220627 Expand Ev Item

+++ app/src/main/java/com/silverback/carman/threads/StationFavRunnable.java
@@ -1,9 +1,10 @@
 package com.silverback.carman.threads;
 
 
-import static com.silverback.carman.threads.FavStationRunnable.RetrofitApi.BASE_URL;
+import static com.silverback.carman.threads.StationFavRunnable.RetrofitApi.BASE_URL;
 
 import android.content.Context;
+import android.net.Uri;
 import android.os.Process;
 
 import androidx.annotation.NonNull;
@@ -14,12 +15,16 @@ import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 import com.silverback.carman.logs.LoggingHelper;
 import com.silverback.carman.logs.LoggingHelperFactory;
-import com.silverback.carman.viewmodels.XmlPullParserHandler;
+import com.silverback.carman.utils.Constants;
 
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.io.InputStream;
+import java.io.ObjectInputStream;
 import java.io.ObjectOutputStream;
+import java.io.Serializable;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -31,9 +36,9 @@ import retrofit2.converter.gson.GsonConverterFactory;
 import retrofit2.http.GET;
 import retrofit2.http.Query;
 
-public class FavStationRunnable implements Runnable {
+public class StationFavRunnable implements Runnable {
 
-    private static final LoggingHelper log = LoggingHelperFactory.create(FavStationRunnable.class);
+    private static final LoggingHelper log = LoggingHelperFactory.create(StationFavRunnable.class);
 
     //private static final String API_KEY = "F186170711";
     //private static final String OPINET = "https://www.opinet.co.kr/api/";
@@ -42,7 +47,8 @@ public class FavStationRunnable implements Runnable {
     // Objects
     private final Context mContext;
     private final StationPriceMethods mCallback;
-    private final XmlPullParserHandler xmlHandler;
+    private String stnId;
+    //private final XmlPullParserHandler xmlHandler;
 
     // Interface
     interface StationPriceMethods {
@@ -50,14 +56,15 @@ public class FavStationRunnable implements Runnable {
         boolean getIsFirst();//if true, the station will be viewed in MainActivity w/ the price.
         void setStnPriceThread(Thread thread);
         void setFavoritePrice(Map<String, Float> data);
+        void setFavStationInfo(Info info);
         void savePriceDiff();
     }
 
     // Constructor
-    FavStationRunnable(Context context, StationPriceMethods callback) {
+    StationFavRunnable(Context context, StationPriceMethods callback) {
         mContext = context;
         mCallback = callback;
-        xmlHandler = new XmlPullParserHandler();
+        //xmlHandler = new XmlPullParserHandler();
     }
 
     @Override
@@ -65,18 +72,29 @@ public class FavStationRunnable implements Runnable {
         android.os.Process.setThreadPriority(Process.THREAD_PRIORITY_BACKGROUND);
         mCallback.setStnPriceThread(Thread.currentThread());
 
-        final String stationId = mCallback.getStationId();
+        stnId = mCallback.getStationId();
         try {
             if(Thread.interrupted()) throw new InterruptedException();
+
+            // Fetch the first favorite station, the price info of which is shown in the price
+            // viewpager in the main activity. Otherwise, just fetch the price info and show
+            // it in the gas manager fragment in the expense activity.
             if(mCallback.getIsFirst()) {
                 Call<FavStationModel> call = RetrofitClient.getIntance()
                         .getRetrofitApi()
-                        .getStationInfoModel("F186170711", stationId, "json");
+                        .getStationInfoModel("F186170711", stnId, "json");
                 call.enqueue(new Callback<FavStationModel>() {
                     @Override
                     public void onResponse(@NonNull Call<FavStationModel> call,
                                            @NonNull Response<FavStationModel> response) {
 
+                        FavStationModel model = response.body();
+                        assert model != null;
+
+                        Info info = model.result.info.get(0);
+                        //List<OilPrice> oilPriceList = model.result.info.get(0).oilPriceList;
+                        calculatePriceDiff(info);
+                        mCallback.setFavStationInfo(info);
                     }
 
                     @Override
@@ -86,12 +104,10 @@ public class FavStationRunnable implements Runnable {
                     }
                 });
 
+
             } else {
 
             }
-
-
-
             /*
             URL url = new URL(URLStn + stationId);
             HttpURLConnection conn = (HttpURLConnection)url.openConnection();
@@ -190,39 +206,107 @@ public class FavStationRunnable implements Runnable {
         public Result result;
     }
 
-    public static class Result {
+    private static class Result {
         @SerializedName("OIL")
         @Expose
-        public List<StationInfoRunnable.Info> info;
+        public List<Info> info;
 
     }
 
-    public static class Info {
+    public static class Info implements Serializable {
+        @SerializedName("UNI_ID")
+        private String stationid;
+        @SerializedName("OS_NM")
+        private String stationName;
+        @SerializedName("POLL_DIV_CD")
+        private String company;
         @SerializedName("VAN_ADR")
         @Expose
         private String addrsOld;
-
         @SerializedName("NEW_ADR")
         @Expose
         private String addrsNew;
-
         @SerializedName("CAR_WASH_YN")
         private String carWashYN;
-
         @SerializedName("CVS_YN")
         private String cvsYN;
-
         @SerializedName("MAINT_YN")
         private String maintYN;
+        @SerializedName("OIL_PRICE")
+        private List<OilPrice> oilPriceList;
+
+        @Expose
+        private Map<String, Float> priceDiff;
+
+        public String getStationid() { return stationid; }
+        public String getStationName() { return stationName; }
+        public String getCompany() { return company; }
+        public List<OilPrice> getOliPriceList() { return oilPriceList; }
+
+
+        public void setPriceDiff(Map<String, Float> priceDiff) {
+            this.priceDiff = priceDiff;
+        }
+        public Map<String, Float> getPriceDiff() {
+            return priceDiff;
+        }
     }
 
-    private void savePriceInfo(File file, Object obj) {
-        log.i("save price info");
-        try(FileOutputStream fos = new FileOutputStream(file);
-            ObjectOutputStream oos = new ObjectOutputStream(fos)) {
-            oos.writeObject(obj);
-            mCallback.savePriceDiff();
-        } catch (IOException e) { e.printStackTrace(); }
+    public static class OilPrice implements Serializable {
+        @SerializedName("PRODCD")
+        private String oilCode;
+        @SerializedName("PRICE")
+        private float price;
+        @SerializedName("TRADE_DT")
+        private String tradeDate;
+        @SerializedName("TRADE_TM")
+        private String tradeTime;
+
+        public String getOilCode() { return oilCode; }
+        public float getPrice() { return price; }
+        public String getTradeDate() { return tradeDate; }
+        public String getTradeTime() { return tradeTime; }
+
+    }
+
+
+
+    private void calculatePriceDiff(Object obj) {
+        final File file = new File(mContext.getFilesDir(), Constants.FILE_FAVORITE_PRICE);
+        if(!file.exists()) {
+            try(FileOutputStream fos = new FileOutputStream(file);
+                ObjectOutputStream oos = new ObjectOutputStream(fos)) {
+                oos.writeObject(obj);
+            } catch (IOException e) { e.printStackTrace(); }
+
+        } else {
+            Uri uri = Uri.fromFile(file);
+            try(InputStream is = mContext.getContentResolver().openInputStream(uri);
+                ObjectInputStream ois = new ObjectInputStream(is)) {
+                final Info savedInfo = (Info)ois.readObject();
+
+                if(stnId.matches(savedInfo.getStationid())) {
+                    Map<String, Float> priceDiff = new HashMap<>();
+                    Info currentInfo = (Info)obj;
+                    for(OilPrice newInfo : currentInfo.oilPriceList) {
+                        for(OilPrice oldInfo : savedInfo.oilPriceList) {
+                            if(newInfo.oilCode.matches(oldInfo.oilCode)) {
+                               priceDiff.put(newInfo.oilCode, (newInfo.price - oldInfo.price));
+                            }
+                        }
+                    }
+                    currentInfo.setPriceDiff(priceDiff);
+
+                } else {
+                    try(FileOutputStream fos = new FileOutputStream(file);
+                        ObjectOutputStream oos = new ObjectOutputStream(fos)) {
+                        oos.writeObject(obj);
+                    } catch (IOException e) { e.printStackTrace(); }
+                }
+            } catch(IOException | ClassNotFoundException | NullPointerException e) {
+                e.printStackTrace();
+            }
+        }
     }
 
 }
